requirements_production.txt
Django==5.2.5
djangorestframework==3.15.2
drf-spectacular==0.27.2
Pillow==10.4.0
reportlab==4.2.2
openpyxl==3.1.2
django-extensions==3.2.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
